+++ R/migrate-hackery.R
@@ -6,10 +6,41 @@ deploy_current <- function(content) {
 
 # ACLs ----------------------------------------------------
 
-acl_add_self <- function(content) {
-  acl_add_collaborator(content, content$get_connect()$GET("me")$guid)
+#' ACL Add Group
+#' 
+#' Add a group_guid to the content as an owner or viewer
+#' 
+#' @param content The R6 Content object
+#' @param group_guid The group's GUID
+#' @param role One of "owner" or "viewer"
+#' 
+#' @return The R6 content object (for piping)
+#' 
+#' @keywords internal
+acl_add_group <- function(content, group_guid, role) {
+  warn_experimental("acl_add")
+  res <- content$get_connect()$POST(
+    glue::glue("applications/{content$get_content()$guid}/groups"),
+    body = list(
+      app_role = role,
+      guid = group_guid
+    )
+  )
+  
+  return(content)
 }
 
+#' ACL Add User
+#' 
+#' Add a user_guid to the content as an owner or viewer
+#' 
+#' @param content The R6 Content object
+#' @param user_guid The user's GUID
+#' @param role One of "owner" or "viewer"
+#' 
+#' @return The R6 content object (for piping)
+#' 
+#' @keywords internal
 acl_add_user <- function(content, user_guid, role) {
   warn_experimental("acl_add")
   res <- content$get_connect()$POST(
@@ -22,16 +53,19 @@ acl_add_user <- function(content, user_guid, role) {
   return(content)
 }
 
+#' @rdname acl_add_user
 acl_add_collaborator <- function(content, user_guid) {
   acl_add_user(content = content, user_guid = user_guid, role = "owner")
 }
 
 # TODO: Should this be a warning if the user is a collaborator? Will downgrade their permissions
 # TODO: How should this behave if the content does not have access_type: acl?
+#' @rdname acl_add_user
 acl_add_viewer <- function(content, user_guid) {
   acl_add_user(content = content, user_guid = user_guid, role = "viewer")
 }
 
+#' @rdname acl_add_user
 acl_remove_user <- function(content, user_guid) {
   warn_experimental("acl_remove")
   res <- content$get_connect()$DELETE(
@@ -40,10 +74,18 @@ acl_remove_user <- function(content, user_guid) {
   return(content)
 }
 
+#' @rdname acl_add_user
 acl_remove_collaborator <- acl_remove_user
 
+#' @rdname acl_add_user
 acl_remove_viewer <- acl_remove_user
 
+#' @rdname acl_add_user
+acl_add_self <- function(content) {
+  acl_add_collaborator(content, content$get_connect()$GET("me")$guid)
+}
+
+#' @rdname acl_add_user
 acl_remove_self <- function(content) {
   acl_remove_user(content, content$get_connect()$GET("me")$guid)
 }
@@ -51,11 +93,32 @@ acl_remove_self <- function(content) {
 acl_user_role <- function(content, user_guid) {
   warn_experimental("acl_user_role")
   scoped_experimental_silence()
-  acls <- get_acl_impl(content)
+  acls <- get_acl_user_impl(content)
   if (is.null(user_guid) || is.na(user_guid)) {
     return(NULL)
   }
   user_entry <- purrr::flatten(purrr::keep(acls, ~ .x$guid == user_guid))
 
   return(user_entry$app_role)
+}
+
+#' @rdname acl_add_group
+acl_remove_group <- function(content, group_guid) {
+  warn_experimental("acl_remove")
+  res <- content$get_connect()$DELETE(
+    glue::glue("applications/{content$get_content()$guid}/groups/{group_guid}")
+  )
+  return(content)
+}
+
+acl_group_role <- function(content, group_guid) {
+  warn_experimental("acl_group_role")
+  scoped_experimental_silence()
+  acls <- get_acl_group_impl(content)
+  if (is.null(group_guid) || is.na(group_guid)) {
+    return(NULL)
+  }
+  group_entry <- purrr::flatten(purrr::keep(acls, ~ .x$guid == group_guid))
+  
+  return(group_entry$app_role)
 }