+++ tests/integrated/test-content.R
@@ -10,10 +10,16 @@ cont1_guid <- NULL
 cont1_bundle <- NULL
 cont1_content <- NULL
 
+cont2_name <- uuid::UUIDgenerate()
+cont2_title <- "Test Content 2"
+
 collab_guid <- NULL
 collab_alt_guid <- NULL
 viewer_guid <- NULL
 viewer_alt_guid <- NULL
+
+collap_group_guid <- NULL
+viewer_group_guid <- NULL
 
 # deploy content
 dir_path <- rprojroot::find_package_root_file("tests/testthat/examples/static")
@@ -24,6 +30,11 @@ tsk <- deploy(connect = test_conn_1, bundle = bund, name = cont1_name, title = c
 
 cont1_guid <- tsk$get_content()$guid
 cont1_content <- content_item(tsk$get_connect(), cont1_guid)
+
+tsk2 <- deploy(connect = test_conn_1, bundle = bund, name = cont2_name, title = cont2_title)
+
+cont2_guid <- tsk2$get_content()$guid
+cont2_content <- content_item(tsk2$get_connect(), cont2_guid)
 
 test_that("content_item works", {
   cont1_tmp <- test_conn_1 %>% content_item(guid = cont1_guid)
@@ -53,11 +64,13 @@ test_that("content_title handles NULL titles gracefully", {
   expect_identical(null_title, "Test Title")
 })
 
+context("acl")
+
 test_that("acl returns owner once and only once", {
   scoped_experimental_silence()
 
   # get acl
-  acls <- get_acl(cont1_content)
+  acls <- get_acl_user(cont1_content)
 
   my_guid <- test_conn_1$GET("me")$guid
 
@@ -89,7 +102,7 @@ test_that("add collaborator twice works", {
   invisible(acl_add_collaborator(cont1_content, collab_guid))
 
   # get acl
-  acls <- get_acl(cont1_content)
+  acls <- get_acl_user(cont1_content)
 
   which_match <- purrr::map2_lgl(acls$guid, acls$app_role, function(.x, .y) {
     .x == collab_guid && .y == "owner"
@@ -108,7 +121,7 @@ test_that("add a viewer works", {
   invisible(acl_add_viewer(cont1_content, viewer_guid))
 
   # get acl
-  acls <- get_acl(cont1_content)
+  acls <- get_acl_user(cont1_content)
 
   which_match <- purrr::map2_lgl(acls$guid, acls$app_role, function(.x, .y) {
     .x == viewer_guid && .y == "viewer"
@@ -124,7 +137,7 @@ test_that("add a viewer twice works", {
   invisible(acl_add_viewer(cont1_content, viewer_guid))
 
   # get acl
-  acls <- get_acl(cont1_content)
+  acls <- get_acl_user(cont1_content)
 
   which_match <- purrr::map2_lgl(acls$guid, acls$app_role, function(.x, .y) {
     .x == viewer_guid && .y == "viewer"
@@ -139,7 +152,7 @@ test_that("remove a collaborator works", {
   invisible(acl_remove_collaborator(cont1_content, collab_guid))
 
   # get acl
-  acls <- get_acl(cont1_content)
+  acls <- get_acl_user(cont1_content)
 
   which_match <- purrr::map2_lgl(acls$guid, acls$app_role, function(.x, .y) {
     .x == collab_guid && .y == "owner"
@@ -154,7 +167,7 @@ test_that("remove a collaborator twice works", {
   invisible(acl_remove_collaborator(cont1_content, collab_guid))
 
   # get acl
-  acls <- get_acl(cont1_content)
+  acls <- get_acl_user(cont1_content)
 
   which_match <- purrr::map2_lgl(acls$guid, acls$app_role, function(.x, .y) {
     .x == collab_guid && .y == "owner"
@@ -173,7 +186,7 @@ test_that("a collaborator does not affect other collaborators", {
   invisible(acl_add_collaborator(cont1_content, collab_guid))
   invisible(acl_add_collaborator(cont1_content, collab_alt_guid))
 
-  acls <- get_acl(cont1_content)
+  acls <- get_acl_user(cont1_content)
 
   # both present
   expect_true(all(c(collab_guid, collab_alt_guid) %in% acls$guid))
@@ -181,7 +194,7 @@ test_that("a collaborator does not affect other collaborators", {
   # remove one
   invisible(acl_remove_collaborator(cont1_content, collab_alt_guid))
 
-  acls2 <- get_acl(cont1_content)
+  acls2 <- get_acl_user(cont1_content)
   # other present
   expect_true(collab_guid %in% acls2$guid)
 })
@@ -192,7 +205,7 @@ test_that("a collaborator and a viewer do not affect each other", {
   invisible(acl_add_collaborator(cont1_content, collab_guid))
   invisible(acl_add_viewer(cont1_content, viewer_guid))
 
-  acls <- get_acl(cont1_content)
+  acls <- get_acl_user(cont1_content)
 
   which_match_collab <- purrr::map2_lgl(acls$guid, acls$app_role, function(.x, .y) {
     .x == collab_guid && .y == "owner"
@@ -204,7 +217,7 @@ test_that("a collaborator and a viewer do not affect each other", {
   expect_true(any(which_match_viewer))
 
   invisible(acl_remove_collaborator(cont1_content, collab_guid))
-  acls2 <- get_acl(cont1_content)
+  acls2 <- get_acl_user(cont1_content)
 
   which_match_viewer2 <- purrr::map2_lgl(acls2$guid, acls2$app_role, function(.x, .y) {
     .x == viewer_guid && .y == "viewer"
@@ -214,7 +227,7 @@ test_that("a collaborator and a viewer do not affect each other", {
   invisible(acl_add_collaborator(cont1_content, collab_guid))
   invisible(acl_remove_viewer(cont1_content, viewer_guid))
 
-  acls3 <- get_acl(cont1_content)
+  acls3 <- get_acl_user(cont1_content)
   which_match_collab3 <- purrr::map2_lgl(acls3$guid, acls3$app_role, function(.x, .y) {
     .x == collab_guid && .y == "owner"
   })
@@ -231,7 +244,7 @@ test_that("a viewer does not affect other viewers", {
   invisible(acl_add_viewer(cont1_content, viewer_guid))
   invisible(acl_add_viewer(cont1_content, viewer_alt_guid))
 
-  acls <- get_acl(cont1_content)
+  acls <- get_acl_user(cont1_content)
 
   # both present
   expect_true(all(c(viewer_guid, viewer_alt_guid) %in% acls$guid))
@@ -239,7 +252,7 @@ test_that("a viewer does not affect other viewers", {
   # remove one
   invisible(acl_remove_viewer(cont1_content, viewer_alt_guid))
 
-  acls2 <- get_acl(cont1_content)
+  acls2 <- get_acl_user(cont1_content)
 
   # other present
   expect_true(viewer_guid %in% acls2$guid)
@@ -256,7 +269,7 @@ test_that("a collaborator can be added as a viewer (overwrites)", {
   # add as viewer
   invisible(acl_add_viewer(cont1_content, collab_guid))
 
-  acls <- get_acl(cont1_content)
+  acls <- get_acl_user(cont1_content)
 
   # TODO: Should this be a warning?
   which_match <- purrr::map2_lgl(acls$guid, acls$app_role, function(.x, .y) {
@@ -276,7 +289,7 @@ test_that("a viewer can be added as a collaborator", {
   # add collaborator
   invisible(acl_add_collaborator(cont1_content, collab_guid))
 
-  acls <- get_acl(cont1_content)
+  acls <- get_acl_user(cont1_content)
 
   which_match <- purrr::map2_lgl(acls$guid, acls$app_role, function(.x, .y) {
     .x == collab_guid && .y == "owner"
@@ -291,7 +304,7 @@ test_that("remove a viewer works", {
   invisible(acl_remove_viewer(cont1_content, viewer_guid))
 
   # get acl
-  acls <- get_acl(cont1_content)
+  acls <- get_acl_user(cont1_content)
 
   which_match <- purrr::map2_lgl(acls$guid, acls$app_role, function(.x, .y) {
     .x == viewer_guid && .y == "viewer"
@@ -306,7 +319,7 @@ test_that("remove a viewer twice works", {
   invisible(acl_remove_viewer(cont1_content, viewer_guid))
 
   # get acl
-  acls <- get_acl(cont1_content)
+  acls <- get_acl_user(cont1_content)
 
   which_match <- purrr::map2_lgl(acls$guid, acls$app_role, function(.x, .y) {
     .x == viewer_guid && .y == "viewer"
@@ -344,4 +357,37 @@ test_that("acl_add_self works", {
 
 test_that("acl_remove_self works", {
   skip("not yet tested")
+})
+
+test_that("acl_add_group works", {
+  scoped_experimental_silence()
+  grp <- test_conn_1$groups_create(name = random_name())
+  
+  content_v1 <- acl_add_group(cont2_content, grp$guid, "owner")
+  
+  expect_is(content_v1, "Content")
+  
+  cacl <- get_acl_group(content_v1)
+  expect_equal(purrr::map_chr(vctrs::vec_ptype(cacl), typeof), purrr::map_chr(vctrs::vec_ptype(connectapi_ptypes$acl_group), typeof))
+  expect_equal(nrow(cacl), 1)
+  expect_equal(acl_group_role(content_v1, grp$guid), "owner")
+  
+  # remove ACL
+  content_v2 <- acl_remove_group(content_v1, grp$guid)
+  cacl_new <- get_acl_group(content_v2)
+  expect_equal(purrr::map_chr(vctrs::vec_ptype(cacl_new), typeof), purrr::map_chr(vctrs::vec_ptype(connectapi_ptypes$acl_group), typeof))
+  expect_equal(nrow(cacl_new), 0)
+  
+  expect_null(acl_group_role(content_v2, grp$guid))
+})
+
+test_that("acl_group_role with null user_guid returns NULL", {
+  scoped_experimental_silence()
+  expect_null(acl_group_role(cont1_content, NULL))
+})
+
+test_that("acl_group_role with no role returns NULL", {
+  scoped_experimental_silence()
+  acl_remove_user(cont1_content, viewer_guid)
+  expect_null(acl_group_role(cont1_content, viewer_guid))
 })